+++ lib/view/course_operations_homeview.dart
@@ -1,10 +1,10 @@
 import 'package:firebase_realtime_usage/core/models/course.dart';
 import 'package:firebase_realtime_usage/core/services/firebase_service.dart';
 import 'package:firebase_realtime_usage/view/course_form_view.dart';
+import 'package:firebase_realtime_usage/view/shared/shared_app_data.dart';
 import 'package:flutter/material.dart';
 
 class CourseHomePage extends StatefulWidget {
-  static int numberOfCourses;
   @override
   _FireHomeViewState createState() => _FireHomeViewState();
 }
@@ -40,7 +40,6 @@ class _FireHomeViewState extends State<CourseHomePage> {
   );
 
    void refreshList() {
-    // reload
     setState(() {
       _courseList = FirebaseService.prefInstance.getCourses();
     });
@@ -51,7 +50,7 @@ class _FireHomeViewState extends State<CourseHomePage> {
       builder: (context, snapshot){
     //whatever returns from this function, will be avaliable inside snapshot paremeter.
     final List<Course> courseList = snapshot.data;
-    CourseHomePage.numberOfCourses =courseList.length;
+    SharedAppData.numberOfCourses =courseList.length;
     switch (snapshot.connectionState) {
       case ConnectionState.waiting:
         {
@@ -81,11 +80,11 @@ class _FireHomeViewState extends State<CourseHomePage> {
 
   Card _courseCard(List<Course> courseList, int index) {
     return Card(
-                  elevation: 4,
-                  child: ListTile(
-                    title: Text(courseList.elementAt(index).name),
-                    subtitle: Text(courseList.elementAt(index).grade),
-                  ),
-                );
+      elevation: 4,
+      child: ListTile(
+        title: Text(courseList.elementAt(index).name),
+        subtitle: Text(courseList.elementAt(index).grade),
+      ),
+    );
   }
 }
