+++ lib/view/shared/shared_app_data.dart
@@ -1,0 +1,3 @@
+class SharedAppData{
+  static int numberOfCourses = 0;
+}
